+++ changelog.md
@@ -8,6 +8,14 @@
     - "Nhom 1": 1 layer
     - "Nhom 2": 2 layers
     - "Nhom 3": 0 layer
+- ks_connection.sde: test connection to local database
+- add libs
+    - listing_layer
+        - get feature layer in mxd 
+    - connect sde
+        - make sde file from connect information
+    - feature2database
 #### remove
 
 #### change
+- change data source in example data